+++ wikiteam3/dumpgenerator/index_php.py
@@ -1,8 +1,7 @@
-import delay
 import os
 
-from clean_html import removeIP
-
+from .delay import delay
+from .util import removeIP
 
 def saveIndexPHP(config={}, session=None):
     """Save index.php as .html, to preserve license details available at the botom of the page"""
